video_factory/requirements.txt
# GUI
PyQt6>=6.6.0

# API клиенты
groq>=0.4.0
google-api-python-client>=2.100.0
elevenlabs>=0.2.0

# Видео обработка
moviepy>=1.0.3
Pillow>=10.0.0

# Браузер автоматизация
selenium>=4.15.0
webdriver-manager>=4.0.0

# Утилиты
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
